Changed files.

+++ ft_bzero.c
@@ -1,14 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_bzero.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: matevos <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/11 13:30:59 by matevos           #+#    #+#             */
+/*   Updated: 2024/01/12 19:57:44 by matevos          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-void    ft_bzero(void * s , int n)
+void	ft_bzero(void *s, size_t n)
 {
-        char    *ptr;
-        int     i;
+	char	*ptr;
+	size_t	i;
 
-        ptr = (char *) s;
-        i = 0;
-        while (i < n)
-        {
-                ptr[i] = 0;                
-        }
+	ptr = (char *) s;
+	i = 0;
+	while (i < n)
+	{
+		ptr[i] = 0;
+	}
 }
